requirements.txt
Django~=5.1.2
psycopg[binary]~=3.2.3
django-environ~=0.11.2
python-dotenv~=1.0.1
gunicorn~=23.0.0
django-extensions
